server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Булат Хазиев <khazievbulatphanzilovich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
django = "^3.1.6"
psycopg2-binary = "^2.8.6"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.2"
SQLAlchemy = "^1.3.23"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
